--- src/components/Form.ts
@@ -1,19 +1,0 @@
-import styled from 'styled-components';
-
-export const Label = styled.label`
-    display: block;
-    margin-bottom: 0.5rem;
-    margin-top: 1.5rem;
-`;
-
-export const Input = styled.input`
-    font-size: 1rem;
-    font-family: inherit;
-    border: none;
-    border-radius: 8px;
-    padding: 0.5rem 0.75rem;
-    box-shadow: 0 0px 1px hsla(0, 0%, 0%, 0.2), 0 1px 2px hsla(0, 0%, 0%, 0.2);
-    background-color: white;
-    line-height: 1.5;
-    margin: 0;
-`;
